moved AbstractRepositoryTest to allow depending packages to use it
