+++ src/State/Testing/AbstractRepositoryTest.php
@@ -9,9 +9,10 @@
  * file that was distributed with this source code.
  */
 
-namespace Broadway\Saga\State;
+namespace Broadway\Saga\State\Testing;
 
 use Broadway\Saga\State;
+use Broadway\Saga\State\Criteria;
 use Broadway\Saga\TestCase;
 
 abstract class AbstractRepositoryTest extends TestCase
@@ -98,7 +99,7 @@ abstract class AbstractRepositoryTest extends TestCase
 
     /**
      * @test
-     * @expectedException Broadway\Saga\State\RepositoryException
+     * @expectedException \Broadway\Saga\State\RepositoryException
      * @expectedExceptionMessage Multiple saga state instances found.
      */
     public function it_throws_an_exception_if_multiple_matching_elements_are_found()
